+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed startup error when user enters a corrupt connection string for platform database locations.  This bug affected syntactically invalid (malformed) connection strings (i.e. not simply connection strings that point to non existant databases)
+- Fixed various issues in ColumnSwapper
+  - If input table contains nulls these are now passed through unchanged
+  - If mapping table contains nulls these are ignored (and not used to map input nulls)
+  - If input table column is of a different Type than the database table a suitable Type conversion is applied
 
 ### Changed